부산대 BE 송승우 1단계 - 엔티티 매핑 (#158)
* docs: 1단계 구현할 기능 목록 추가

* feat: Member 엔티티 클래스 작성

* feat: Wish 엔티티 클래스 작성

* refactor: domain 패키지 내 Wish, Member, Product 클래스를 @Entity로 설정

* refactor: MemberJdbcRepository를 JPA 기반 MemberRepository로 전환

* refactor: ProductJdbcRepository를 JPA 기반 ProductRepository로 전환

* refactor: WishListJdbcRepository를 JPA 기반 WishListRepository로 전환

* feat: @DataJpaTest를 활용한 레포지토리 테스트 코드 작성

+++ src/main/java/gift/domain/WishSummary.java
@@ -2,14 +2,14 @@ package gift.domain;
 
 public class WishSummary {
     private String productName;
-    private int count;
+    private Long count;
 
-    public WishSummary(String productName, int count) {
+    public WishSummary(String productName, Long count) {
         this.productName = productName;
         this.count = count;
     }
 
-    public static WishSummary of(String productName, int count) {
+    public static WishSummary of(String productName, Long count) {
         return new WishSummary(productName, count);
     }
 
@@ -17,7 +17,7 @@ public class WishSummary {
         return productName;
     }
 
-    public int getCount() {
+    public Long getCount() {
         return count;
     }
 }
